[skip ci] move script to alpha
